chore: :wrench: revert pnpm replace with bun
