--- bunfig.toml
@@ -1,8 +1,0 @@
-[install.lockfile]
-
-# whether to save the lockfile to disk
-save = true
-
-# whether to save a non-Bun lockfile alongside bun.lockb
-# only "yarn" is supported
-print = "yarn"
